debian/testsuite-xfail-debian.mk: turn xfail on for failing tests
Lot of tests has already been set to xfail are running inside a
container and would like to mounting /proc with no success.
The affected tests in this patch have the same problem.

In our case the build is running inside a container, that causes
this mentioned tests are running in a 'container-in-container'
environment which usually needs a special preparation and operations in
a strict order to make working everything.